refactor(db): move testFactories to db/utils
Relocate testFactories from db/repositories to db/utils to centralize
test utilities. Update all import paths in collectionsRepository,
settingsRepository, and trackRepository test files accordingly.
